Import samsung LL kernel source
Conflicts:
	Makefile
	arch/arm/boot/compressed/Makefile
	arch/arm/configs/chagalllte_00_defconfig
	arch/arm/configs/ha3g_00_defconfig
	arch/arm/configs/ha3g_03_defconfig
	arch/arm/configs/lt033g_00_defconfig
	arch/arm/configs/smdk5420_defconfig
	arch/arm/configs/universal5410_defconfig
	arch/arm/mach-exynos/board-klimt-gpio.c
	arch/arm/mach-exynos/board-lt03-battery.c
	arch/arm/mach-exynos/sec_misc.c
	drivers/gpu/arm/mali400/ump_r3p0/include/ump_kernel_interface.h
	drivers/gpu/arm/mali400/ump_r3p0/include/ump_kernel_interface_ref_drv.h
	drivers/gpu/arm/mali400/ump_r3p0/include/ump_kernel_interface_vcm.h
	drivers/gpu/arm/mali400/ump_r3p0/include/ump_kernel_platform.h
	drivers/gpu/ion/ion.c
	drivers/hid/hid-core.c
	drivers/hid/hid-pl.c
	drivers/ice4_fpga/ice4_fpga.c
	drivers/md/dm-crypt.c
	drivers/media/video/exynos/mfc/s5p_mfc.c
	drivers/thermal/exynos_thermal.c
	fs/namespace.c
	include/linux/rtnetlink.h
	include/net/fib_rules.h
	include/net/flow.h
	include/net/route.h
	kernel/futex.c
	net/bluetooth/hci_conn.c
	net/core/fib_rules.c
	net/ipv4/fib_frontend.c
	net/ipv4/raw.c
	net/ipv4/route.c
	net/ipv6/ip6_output.c
	net/ipv6/route.c
	net/sctp/sm_make_chunk.c
	security/keys/gc.c
	sound/core/control.c

+++ fs/sdcardfs/packagelist.c
@@ -70,7 +70,7 @@ static int contain_appid_key(struct packagelist_data *pkgl_dat, void *appid) {
         struct hashtable_entry *hash_cur;
 	struct hlist_node *h_n;
 
-        hash_for_each_possible(pkgl_dat->appid_with_rw,	hash_cur, hlist, (unsigned int)appid, h_n)
+        hash_for_each_possible(pkgl_dat->appid_with_rw,	hash_cur, hlist, (uintptr_t)appid, h_n)
                 if (appid == hash_cur->key)
                         return 1;
 	return 0;
@@ -89,7 +89,7 @@ int get_caller_has_rw_locked(void *pkgl_id, derive_t derive) {
 
 	appid = multiuser_get_app_id(current_fsuid());
 	mutex_lock(&pkgl_dat->hashtable_lock);
-	ret = contain_appid_key(pkgl_dat, (void *)appid);
+	ret = contain_appid_key(pkgl_dat, (void *)(uintptr_t)appid);
 	mutex_unlock(&pkgl_dat->hashtable_lock);
 	return ret;
 }
@@ -206,7 +206,7 @@ static int insert_int_to_null(struct packagelist_data *pkgl_dat, void *key, int 
 
 	//printk(KERN_INFO "sdcardfs: %s: %d: %d\n", __func__, (int)key, value);
 	hash_for_each_possible(pkgl_dat->appid_with_rw,	hash_cur, hlist,
-					(unsigned int)key, h_n) {
+					(uintptr_t)key, h_n) {
 		if (key == hash_cur->key) {
 			hash_cur->value = value;
 			return 0;
@@ -218,7 +218,7 @@ static int insert_int_to_null(struct packagelist_data *pkgl_dat, void *key, int 
 	new_entry->key = key;
 	new_entry->value = value;
 	hash_add(pkgl_dat->appid_with_rw, &new_entry->hlist,
-			(unsigned int)new_entry->key);
+			(uintptr_t)new_entry->key);
 	return 0;
 }
 
@@ -294,7 +294,7 @@ static int read_package_list(struct packagelist_data *pkgl_dat) {
 			while (token != NULL) {
 				if (!kstrtoul(token, 10, &ret_gid) &&
 						(ret_gid == pkgl_dat->write_gid)) {
-					ret = insert_int_to_null(pkgl_dat, (void *)appid, 1);
+					ret = insert_int_to_null(pkgl_dat, (void *)(uintptr_t)appid, 1);
 					if (ret) {
 						sys_close(fd);
 						mutex_unlock(&pkgl_dat->hashtable_lock);
